requirements.txt
PySide6
PyPDF2==2.10.0
PyMuPDF
Pillow